@@ -5,10 +5,11 @@ author: bytebeats
 homepage: https://github.com/bytebeats/marquee_flutter
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
+  fading_edge_scrollview: ^2.0.0
   flutter:
     sdk: flutter
 
