@@ -1,4 +1,6 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:marquee_flutter/marquee_text.dart';
 
 void main() {
   runApp(MyApp());
@@ -9,7 +11,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
+      title: 'Marquee Text',
       theme: ThemeData(
         // This is the theme of your application.
         //
@@ -21,18 +23,14 @@ class MyApp extends StatelessWidget {
         // Notice that the counter didn't reset back to zero; the application
         // is not restarted.
         primarySwatch: Colors.blue,
-        // This makes the visual density adapt to the platform that you run
-        // the app on. For desktop platforms, the controls will be smaller and
-        // closer together (more dense) than on mobile platforms.
-        visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
+      home: MyHomePage(title: 'Marquee Text App'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -50,68 +48,75 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      _counter++;
-    });
-  }
+  bool _useRtlText = false;
 
   @override
   Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
-    return Scaffold(
-      appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
-      ),
-      body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Invoke "debug painting" (press "p" in the console, choose the
-          // "Toggle Debug Paint" action from the Flutter Inspector in Android
-          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
-          // to see the wireframe for each widget.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headline4,
-            ),
-          ],
+    return MaterialApp(
+      title: "Marquee Text",
+      home: Scaffold(
+        backgroundColor: Colors.deepPurple,
+        body: ListView(
+          padding: EdgeInsets.only(top: 50),
+          children: [
+            _buildMarqueeText(),
+            _buildComplicateMarqueeText(),
+          ].map((e) => _wrapWithStuff(e)).toList(),
+        ),
+        floatingActionButton: FloatingActionButton.extended(
+          onPressed: () => {},
+          label: !_useRtlText ? const Text(("Switch to Hebrew")) : const Text("العبرية"),
+          backgroundColor: Colors.brown,
         ),
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
+    );
+  }
+
+  Widget _buildMarqueeText() {
+    return MarqueeText(
+      key: Key("$_useRtlText"),
+      text: !_useRtlText
+          ? "This is a Flutter Marquee Text, have a try and spread the word!"
+          : "هذا هو النص مع ترجمة عائمة ، وهناك محاولة لنشر الكلمات !",
+      velocity: 50,
+    );
+  }
+
+  Widget _buildComplicateMarqueeText() {
+    return MarqueeText(
+      text: !_useRtlText
+          ? "There once was a boy who told this story about a boy: "
+          : "هذا هو النص مع ترجمة عائمة ، وهناك محاولة لنشر الكلمات !",
+      style: TextStyle(fontWeight: FontWeight.bold),
+      scrollAxis: Axis.horizontal,
+      crossAxisAlignment: CrossAxisAlignment.start,
+      blankSpace: 15,
+      velocity: 35,
+      pauseAfterRoundFinish: Duration(milliseconds: 800),
+      showFadingEdgeOnlyWhenScrolling: true,
+      fadingEdgeStartFraction: 0.2,
+      fadingEdgeEndFraction: 0.2,
+      roundCounts: 5,
+      paddingAfterText: 20,
+      acceleratingDuration: Duration(milliseconds: 500),
+      acceleratingCurve: Curves.linear,
+      deceleratingDuration: Duration(milliseconds: 500),
+      deceleratingCurve: Curves.easeOut,
+      textScaleFactor: 0.15,
+      textDirection: !_useRtlText ? TextDirection.ltr : TextDirection.rtl,
+      startAfter: Duration(milliseconds: 300),
+      onDone: () => {debugPrint("round done!")},
+    );
+  }
+
+  Widget _wrapWithStuff(Widget child) {
+    return Padding(
+      padding: EdgeInsets.all(20.0),
+      child: Container(
+        height: 40,
+        color: Colors.deepOrange,
+        child: child,
+      ),
     );
   }
 }
